Change preco to BigDecimal

+++ src/main/java/com/delegrego/api_produtos/dto/ProdutoListResponse.java
@@ -1,12 +1,14 @@
 package com.delegrego.api_produtos.dto;
 
+import java.math.BigDecimal;
+
 public record ProdutoListResponse(
 
 		int id,
 
 		String nome,
 
-		double preco,
+		BigDecimal preco,
 
 		String urlImagem
